% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windConversion.R
\name{windConversion}
\alias{windConversion}
\title{Wind speed unit conversion between knots and mph}
\usage{
windConversion(wind_speed, units)
}
\arguments{
\item{wind_speed}{Wind speed}

\item{units}{Specify if wind speed entered is in "knots" or "mph"}
}
\description{
Wind speed unit conversion between knots and mph
}
\examples{
windConversion(15, "mph")
}
